% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ngram.R
\name{add_ngram}
\alias{add_ngram}
\alias{add_ngram.default}
\alias{add_ngram.mixdb}
\title{Add ngram to a mixdb}
\usage{
add_ngram(x, n_gram = 2L)

\method{add_ngram}{default}(x, n_gram = 2L)

\method{add_ngram}{mixdb}(x, n_gram = 2L)
}
\arguments{
\item{x}{a \code{\link{mixdb}} object.}

\item{n_gram}{(int, default = 2) maximum level of ngram to consider
to enrich \code{\link{mixdb}}'s texts and dictionary.}
}
\value{
a \code{\link{mixdb}} wiht the ngram added
}
\description{
Add ngram to a mixdb
}
\examples{
library(tibble)
library(limpido)
test <- tribble(
  ~class, ~id, ~notes,                  ~gender,
     0  ,   1, "foo notes foo",          "male",
     0  ,   2, "bar notes foo",          "female",
     1  ,   1, "another notes",          "male",
     1  ,   3, "annotated foo one two",  "female"
)
foo <- mixdb(test, meta_vars(id, gender))
add_ngram(foo)
add_ngram(foo, 3)
}
