% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_oov_when_grater_than.R
\name{add_oov_when_greater_than}
\alias{add_oov_when_greater_than}
\title{Add \strong{OOV} if out of vocabulary}
\usage{
add_oov_when_greater_than(.list, max_words)
}
\arguments{
\item{.list}{(list) tokens, i.e. named integers (each integer is the
name's index in the dictionary), to considered}

\item{max_words}{(int) maximum index considered in the current
representation}
}
\value{
a modified version of \code{.list} with index \code{max_word + 1}, named "\strong{OOV}", for all the indeces exceeding \code{max_word}
}
\description{
Convert tokens inside \code{.list} in a "\strong{OOV}" if their representation
as an integer in the dictionary is greater than the maximum allowed
one as set in \code{max_word}
}
\examples{

corpus <- list(
  a = setNames(seq_along(letters), letters),
  A = setNames(seq_along(letters), LETTERS)[1:10]
)
add_oov_when_greater_than(corpus, 20)

}
