% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_dict.R
\name{c_dict}
\alias{c_dict}
\title{Combine dictionaries}
\usage{
c_dict(dict_1, dict_2)
}
\arguments{
\item{dict_1}{a \code{\link{dictionary}}}

\item{dict_2}{a \code{\link{dictionary}}}
}
\value{
a \code{\link{dictionary}}
}
\description{
Combine dictionaries
}
\examples{
dict_1 <- dictionary(c("a", "b", "b"))
dict_2 <- dictionary(c("a", "b", "c", "c", "c", "c"))
c_dict(dict_1, dict_2)
}
\seealso{
\code{\link{dictionary}}
}
