% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_num.R
\name{code_num}
\alias{code_num}
\alias{code_num.character}
\title{Code numbers}
\usage{
code_num(text, code, ignore_in_word)

\method{code_num}{character}(text, code = "[NUM]",
  ignore_in_word = FALSE)
}
\arguments{
\item{text}{(chr) vector of text to in which code the numbers}

\item{code}{(chr) code to use to substitute the numbers}
}
\value{
(chr) vector of the same length of \code{text} with al the numbers
changed in \code{code}
}
\description{
Code numbers
}
\examples{
code_num(c("1", "12", "abc12", "abc.12", "l'82"))
}
