% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ngram.R
\name{create_ngram}
\alias{create_ngram}
\alias{create_ngram.numeric}
\alias{create_ngram.character}
\title{Create ngram set}
\usage{
create_ngram(x, n_gram = 2L)

\method{create_ngram}{numeric}(x, n_gram = 2)

\method{create_ngram}{character}(x, n_gram = 2L)
}
\arguments{
\item{x}{an object of class \code{\link{mixdb}}.}

\item{n_gram}{(int, default = 2L) number ($n$) of gram to
create.}
}
\value{
(table) named sorted table of
}
\description{
Create ngram set
}
