% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_otiti_mixdb.R
\name{create_otiti_mixdb}
\alias{create_otiti_mixdb}
\title{Create the main mixdb for the otiti project}
\usage{
create_otiti_mixdb(
  data_path = here::here("../data"),
  pedia_gold_path = file.path(data_path, "pedia_gold_otiti.rds"),
  output_path = file.path(data_path, "mixdb_otiti_tagged.rds")
)
}
\arguments{
\item{data_path}{(chr) path to the data folder}

\item{pedia_gold_path}{(chr) path to pedinet gold standard file RDS}

\item{output_path}{(chr) path to store the output mixdb in RDS format}
}
\value{
invisibly `TRUE``
}
\description{
Create the main \link{mixdb} for the otiti project
}
