% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_predict.R
\name{ensemble_predict}
\alias{ensemble_predict}
\title{Ensemble predict}
\usage{
ensemble_predict(models, params)
}
\arguments{
\item{models}{(lst) a list of (unserialized) models}

\item{params}{object params output stored in the training phase}
}
\value{
list of performances
}
\description{
Ensemble predict
}
