% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_punctuations.R
\name{expand_punctuations}
\alias{expand_punctuations}
\alias{expand_punctuations.character}
\title{expand punctuation}
\usage{
expand_punctuations(text, regex)

\method{expand_punctuations}{character}(text, regex = "\\\\W+")
}
\arguments{
\item{text}{(chr) vector}

\item{regex}{(chr) representing the regular expression to use to
match words}
}
\value{
a character vector with all punctuation, sourrendered by
spaces
}
\description{
expand punctuation
}
\examples{
expand_punctuations(c("abc de", "a.b", "A.B", "c'è", "s.p.a."))
}
