% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixdb.R
\name{get_dictionary}
\alias{get_dictionary}
\alias{get_dictionary.mixdb}
\title{Dictionary extractor}
\usage{
get_dictionary(x)

\method{get_dictionary}{mixdb}(x)
}
\arguments{
\item{x}{an object from which extract the \code{\link{dictionary}}}
}
\value{
a \code{\link{dictionary}}
}
\description{
Dictionary extractor
}
\examples{
library(tibble)
library(limpido)
test <- tribble(
  ~class, ~id, ~notes,      ~gender,
     0  ,   1, "foo notes",     "male",
     0  ,   2, "bar notes",     "female",
     1  ,   1, "another notes", "male",
     1  ,   3, "annotated foo", "female"
)
mixdb(test, meta_vars(id, gender)) \%>\%
    get_dictionary()
}
\seealso{
\code{\link{dictionary}}
}
