% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies.R
\name{get_frequencies}
\alias{get_frequencies}
\alias{get_frequencies.dictionary}
\title{Extractor for the attribute \code{frequencies}, of objects of class
\code{\link{dictionary}}}
\usage{
get_frequencies(dict)

\method{get_frequencies}{dictionary}(dict)
}
\arguments{
\item{dict}{a \code{\link{dictionary}} object.}
}
\value{
(int) named vector of the words' frequencies in the
\code{\link{dictionary}}
}
\description{
Extractor for the attribute \code{frequencies}, of objects of class
\code{\link{dictionary}}
}
\examples{
dic <- dictionary(c("a", "b", "b"))
get_frequencies(dic)
}
\seealso{
\code{\link{dictionary}}
}
