% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{get_metrics}
\alias{get_metrics}
\title{Get metrix from trained objects}
\usage{
get_metrics(path = here::here("../output/"))
}
\arguments{
\item{path}{(chr, default = "../output") output path for the stored models}
}
\value{
a tibble
}
\description{
given the path to the output folder in which the \code{.rda} file output
of the script in \code{limpido/analyses/R} are stored, provides useful
metrics from the model
}
\examples{
\dontrun{
    get_metrics()
}
}
