% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_history.R
\name{gg_history}
\alias{gg_history}
\title{Custom ggplot for keras' historys}
\usage{
gg_history(keras_plot, history, architecture, params,
  train_time = lubridate::now())
}
\arguments{
\item{keras_plot}{a keras plot object}

\item{history}{a [keras]{history} object}

\item{architecture}{(chr) a description of the architecture used}

\item{params}{output of \link{setup_input_data}}

\item{train_time}{(POSIXct, dafault \code{lubridate::now()}) time to
include in the plot.}
}
\value{
a [ggplot2]{ggplot} object
}
\description{
Custom ggplot for keras' historys
}
