% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_vars.R
\name{meta_vars}
\alias{meta_vars}
\title{This helper is intended to provide equivalent semantics to
\code{\link[dplyr]{select}()}}
\usage{
meta_vars(...)
}
\arguments{
\item{...}{Variables to include/exclude like meta in the creation
of a \code{\link{mixdb}} object}
}
\description{
This helper is intended to provide equivalent semantics to
\code{\link[dplyr]{select}()}
}
