% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_otiti_classes.R
\name{predict_otiti_classes}
\alias{predict_otiti_classes}
\title{predict otitis classes with metrics}
\usage{
predict_otiti_classes(.model, params)
}
\arguments{
\item{.model}{a Keras model}

\item{params}{a list as output of \link{setup_input_data}}
}
\value{
a list with four slots:
\itemize{
\item vector of validation predictions
\item vactor of validation gold labels
\item list of metrices (crude and balanced accuracy, precision, recall,
and  f1)
\item list of wrong classified information (classes predicted, and gold
ones, and an extraction of wrong records from the full dataset)
}
}
\description{
given a Keras model, and parameters as output of \link{setup_input_data},
it predicts the validation classes and compute some metrics of
interest
}
