% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_db.R
\name{prepare_db}
\alias{prepare_db}
\title{DB preparation}
\usage{
prepare_db(final_pedia_db)
}
\arguments{
\item{final_pedia_db}{dataframe with the final version of
the full pedianet db.}
}
\value{
(chr) Character vector with each element a cell of the
\code{final_pedia_db} preprocessed
}
\description{
perform vectorization of all the pertinent columns in a single
vector of texts.
}
\examples{
\dontrun{
    expanded <- prepare_db(dataset_final)
}
}
