% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_all_models.R
\name{save_all_models}
\alias{save_all_models}
\title{Save all the relevant object for a DL model}
\usage{
save_all_models(model, history, output_path, plot = NULL,
  .now = lubridate::now())
}
\arguments{
\item{model}{a keras model}

\item{history}{akeras history}

\item{output_path}{(chr) the path for the output folder}

\item{plot}{(ggplot, default = NULL) if a plot is provided, it save
a PNG of the plot too}

\item{.now}{(POSIXct, dafault \code{lubridate::now()}) time to include in
the saved files'names.}
}
\value{
invisibly \code{TRUE}
}
\description{
it saves:
}
