% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_no_meta.R
\name{text_no_meta}
\alias{text_no_meta}
\title{Select textual column excluding meta information}
\usage{
text_no_meta(x, meta = NULL)
}
\arguments{
\item{x}{(data frame) input data from which extract only the textual
columns not included in \code{meta}}

\item{meta}{unquoted list of column names passed using the
\code{\link{meta_vars}()} function.}
}
\description{
Select textual column excluding meta information
}
