% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpido-package.R
\docType{package}
\name{limpido-package}
\alias{limpido}
\alias{limpido-package}
\title{limpido: LearnIng Machine for PedIatric Detection of Otitis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com}

}
\keyword{internal}
