% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpido-package.R
\docType{package}
\name{limpido-package}
\alias{limpido}
\alias{limpido-package}
\title{limpido: LearnIng Machine for PedIatric Detection of Otitis}
\description{
The goal of the package `limpido` is to develop a "LearnIng Machine for PedIatric Detection of Otitis".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UBESP-DCTV/limpido}
  \item Report bugs at \url{https://github.com/UBESP-DCTV/limpido/issues}
}

}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@gmail.com}

Other contributors:
\itemize{
  \item Elisa Barbieri \email{elisa.barbieri@unipd.it} [contributor]
  \item Carlo Giacquinto \email{carlo.giaquinto@unipd.it} [contributor]
  \item Dario Gregori \email{dario.gregori@unipd.it} [contributor]
}

}
\keyword{internal}
