% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixdb.R
\name{mixdb}
\alias{mixdb}
\alias{mixdb.default}
\alias{mixdb.data.frame}
\title{mixdb object}
\usage{
mixdb(.data, .meta_vars = NULL)

\method{mixdb}{default}(.data, .meta_vars = NULL)

\method{mixdb}{data.frame}(.data, .meta_vars = NULL)
}
\arguments{
\item{.data}{(data frame) of data. (see \strong{details} for specs)}

\item{.meta_vars}{optional list of (unquoted) columns' names generated
by \code{\link{meta_vars}()}, and representing metadata of
interest. (see \strong{details} for specs)}
}
\value{
an object of class \code{\link[limpido]{mixdb}}, with an
attribute \strong{meta} of class \link[tibble:tibble-package]{tibble}
reporting the additional provided and selected (by \code{.meta_vars})
information.
}
\description{
\code{\link[limpido]{mixdb}} reads a .data general database stored in a
data.frame and convert it in a parsed list for mlt algorithms (keras
and tensorflow) aimed to Natural Language or mixed formad analyses.
}
\details{
The \code{.data} data frame in input must have the following
(named) column:
\describe{
\item{class}{(factor) classification's labels, stored in a
factor. It accepts also an integer (or something cohercible to
it) or character vector, creating a factor from them. Any other
class will return an error.
}
}

All the other columns listed in the \code{.meta_vars} parameter will
represents metadata of the datasets, useful for other purposes then
NLP, e.g., possible future non-textual covariates for the models.

All the character columns not listed in \code{.meta_vars} will be
treated as textual ones and converted, accordingly to the
dictionary provided

All the other columns will be ignored and not saved anywhere.
}
\note{
As a convention, "0" does not stand for a specific word, but
instead is used to encode any unknown word.
}
\examples{
library(tibble)
library(limpido)
test <- tribble(
  ~class, ~id, ~notes,                  ~gender,
     0  ,   1, "foo notes foo",          "male",
     0  ,   2, "bar notes foo",          "female",
     1  ,   1, "another notes",          "male",
     1  ,   3, "annotated foo one two",  "female"
)
mixdb(test, meta_vars(id, gender))
mixdb(test, meta_vars(id))

}
